signal: workaround a printf long vs int type coercion issue.
* ext/posix/signal.c (sig_handle): Some architectures use a long
for sig_atomic_t, so we need a %ld specifier to display it without
a warning.  In case sig_atomic_t is narrower than a long, cast it
up before passing to fprintf.

+++ ext/posix/signal.c
@@ -142,7 +142,7 @@ sig_handle (lua_State *L, lua_Debug *LPOSIX_UNUSED (ar))
 		/* Call handler with signal number */
 		lua_pushinteger(L, signalno);
 		if (lua_pcall(L, 1, 0, 0) != 0)
-			fprintf(stderr,"error in signal handler %d: %s\n",signalno,lua_tostring(L,-1));
+			fprintf(stderr,"error in signal handler %ld: %s\n", (long)signalno, lua_tostring(L,-1));
 	}
 	signal_count = 0;  /* reset global to initial state */
